+++ src/pages/Login.tsx
@@ -4,10 +4,10 @@ import Form from '@components/Form';
 
 const Login: React.FC = (): JSX.Element =>
   <Alignment vertical horizontal>
-    <Form>
-      <Form.Input placeholder='Email Address' />
-      <Form.Input placeholder='Password' password />
-    </Form>
+      <Form onSubmit={ login } >
+        <Form.Input onChange={ handleChange('email') } placeholder='Email Address' />
+        <Form.Input onChange={ handleChange('password') } placeholder='Password' password />
+      </Form>
   </Alignment>;
 
 export default Login;