+++ docs/changelog.md
@@ -43,6 +43,7 @@ Sprint 9 / March 28, 2020
 - added constructStringFromSeconds utility function.
 - updated Timesheet records to use utilities' constructStringFromSeconds function to display records as time strings.
 - added amazon-cognito-identity-js and aws-sdk packages.
+- added onSubmit prop to Form submit input.
 
 ## Release 0.8.0
 Sprint 8 / March 14, 2020